+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Optional dependency: ``wheels_only`` for solvers with pre-compiled binaries
+- ReLUQP solver
 
 ### Changed